Fix UI snags: Create Profile, Just Ride, Begin Class, dark theme toggle, and nav bar padding
- NavGraph: Wired onCreateProfile to show ProfileCreationDialog, fixed class_detail route to show ClassDetailScreen, added hasJustRideStarted state for JustRideScreen
- MainActivity: Added Room DB loading via LaunchedEffect, wrapped content in Surface with MaterialTheme.colorScheme.background
- Theme.kt: Added darkTheme parameter with PelonotLightColorScheme
- Color.kt: Added light theme colors (LightBackground, LightSurface, LightSurfaceVariant, DarkText, DarkTextSecondary)
- SettingsScreen: Wired onThemeToggle callback to propagate theme changes
- All screen files: Replaced hardcoded TextPrimary/TextSecondary with MaterialTheme.colorScheme colors
- JustRideScreen: New screen for free ride mode with placeholder metrics
- PLAN.md: Added items 8.10 (guest post-ride prompt), 8.11 (Material Expressive), 8.12 (light mode fix)

+++ app/src/main/java/com/pelonot/data/sensor/SerialPortReader.kt
@@ -45,8 +45,8 @@ class SerialPortReader {
     private var currentResistance: Double = 0.0
     private var currentCadence: Double = 0.0
     private var reconnectJob: Job? = null
-    private var reconnectAttempts: Long = 0
-    private val maxReconnectAttempts = 10L
+    private var reconnectAttempts: Int = 0
+    private val maxReconnectAttempts = 10
     private val baseDelayMs = 1000L
     
     /**
@@ -200,7 +200,7 @@ class SerialPortReader {
         reconnectJob = scope.launch {
             while (isActive && reconnectAttempts < maxReconnectAttempts) {
                 reconnectAttempts++
-                val delayMs = baseDelayMs * (1L shl (reconnectAttempts - 1).coerceAtMost(5))
+                val delayMs = baseDelayMs * (1 shl (reconnectAttempts - 1).coerceAtMost(5)).toLong()
                 Log.i(TAG, "Attempting serial port reconnect (attempt $reconnectAttempts, delay ${delayMs}ms)")
                 
                 delay(delayMs)
